Move patching methods into Patcher class
Create class Patcher and move patching related methods into it.
